requirements.txt
# MAPLAB Digital Asset Pipeline — Python Dependencies
# Install: pip install -r requirements.txt

# Google APIs
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0

# AI Vision
google-generativeai==0.5.4

# Image Processing
Pillow==10.3.0

# Location / Geocoding
geopy==2.4.1

# Notion
notion-client==2.2.1

# Utilities
python-dotenv==1.0.1
